Add method for explicitly authenticating.
For the Ruby version of the HTTP collector, I needed to know if we had successfully authenticated with the provided credentials. The current version of the agent sorta has this information in a really inconvenient way. The `authenticate!` method starts the connection and waits to see if we authenticate. If we don't in 3 seconds, we call it a failure. Also added an accessor.