Why is this change needed? -------------------------- Move the requeue code back to where it was previously
How does it address the issue?
------------------------------
The code was moved in the first place to requeue even if the agent thread was going to be stopped.  While technically (arguably?) more correct, this is a big behavior change, and it is highly likely that most agents in this state will never be restarted.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------