Why is this change needed? -------------------------- There were some ambiguous metric names and a potential requeue-todo
How does it address the issue?
------------------------------
Change the names to be unique.  Also, just assume that no re-enqueue should take place.  I don't love it, but it doesn't break current behavior, and doesn't require changing other tests or the agent code to be more complex with regard to connection refused and related errors.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------