Introduce EventAggregator to the agent and implement commands as structs.
Why is this change needed?
--------------------------
Aggregation is going to be messy and agent.rb is messy enough. Also, there is a lot of string/time mangling going on throughout that can be simplified.

How does it address the issue?
------------------------------
The EventAggregator class can be more easily tested and is a better abstraction than hash-mashing. Implementing instrumental commands as structs simplifies rules around data consistency and should be somewhat faster than constantly manipulating strings to accomplish aggregation.

Please not that this commit is NOT expected to pass all of the tests.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------