Introduce a seperate thread for aggregation of incoming metrics, and fairly complete tests for the same.
Why is this change needed?
--------------------------
Rather than sending metrics one at a time, we can aggregate them in reasonable windows on the client side, to reduce network load.

How does it address the issue?
------------------------------
Aggregate metrics for a while (1 to 60 seconds) before sending them, and then send them in big batches.

Yet to do:
Tests for when the aggregator is getting backed up (and how to detect that)
What default should we use?  On or off?  6, 10, or 15 seconds?

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------