Update agent default frequency Make cases where frequency = 0 more resilient Make agent specs more explicit about starting at the beginning of a frequency window
Why is this change needed?
--------------------------
Default behavior should be the same as agent 2.1.x, for ease of upgrading
General issues we found while testing revealed the need to be more explicit in agent setup and when sending commands with timing data

How does it address the issue?
------------------------------
By starting time-sensitive tests at the start of a frequency window, we avoid the case where subsequent metrics could be split across frequencies unexpectedly.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------

Co-authored-by: Jason Gladish <jason@expectedbehavior.com>