Fix some timing and threading issues in the tests, and simplify the logic when calling at_exit.
Why is this change needed?
--------------------------
If you do threading exactly wrong, you can get an agent to drop data when it should not. Also, you can get a test expectation that takes longer to calculate than the total time the test should take.

How does it address the issue?
------------------------------
We run an exit-flush when exiting, regardless of the state of the queues. This is very fast, but necessary in the world of multiple background threads and aggregators.

Also, we force rspec to pre-load all its exception diffing and matching and other code, so that it is much faster up front, rather than much slower later.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------

Co-authored-by: Jason Gladish <jason@expectedbehavior.com>